requirements.txt
# Weighted Voronoi Stippling Dependencies
# Python 3.8+ required

# Core dependencies
numpy>=1.21.0
Pillow>=8.0.0
scipy>=1.7.0
matplotlib>=3.3.0

# Performance optimization dependencies
numba>=0.56.0

# GPU acceleration (optional)
# cupy-cuda11x>=10.0.0  # For CUDA 11.x
# cupy-cuda12x>=12.0.0  # For CUDA 12.x
